requirements.txt
aioredis==1.3.1
APScheduler==3.7.0
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bleach==4.1.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
click==8.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
daphne==3.0.2
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2
django-apscheduler==0.6.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-mirage-field==1.2.2
django-redis==5.0.0
django-taggit==1.5.1
djangorestframework==3.12.4
drf-yasg2==1.19.4
elasticsearch==7.14.1
elasticsearch-dsl==7.4.0
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.68.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
inflection==0.5.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itypes==1.2.0
JayDeBeApi==1.2.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.37.0
lxml==4.6.3
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.0
openpyxl==3.0.7
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.6.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydantic==1.8.2
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyquery==1.4.3
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
pytz==2019.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
redis==3.5.3
reportlab==3.6.1
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.10
starlette==0.14.2
systemd-python==234
terminado==0.12.1
testpath==0.5.0
testresources==2.0.1
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
zope.interface==5.4.0
